ruma

+++ config/application.rb
@@ -1,7 +1,12 @@
 require_relative "boot"
-require 'rails/all'
-require 'dotenv/load' if defined?(Rails) && (Rails.env.development? || Rails.env.test?)
+require "rails/all"
 require 'line/bot'
+
+# dotenv の読み込みを開発環境とテスト環境のみに制限
+if Rails.env.development? || Rails.env.test?
+  require 'dotenv/load'
+end
+
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
@@ -16,12 +21,18 @@ module HotelLine
     # Common ones are `templates`, `generators`, or `middleware`, for example.
     config.autoload_lib(ignore: %w[assets tasks])
 
-    config.before_configuration do
-      env_file = File.join(Rails.root, 'config', 'local_env.yml')
-      YAML.load(File.open(env_file)).each do |key, value|
-        ENV[key.to_s] = value
-      end if File.exist?(env_file)
+    # 環境変数の読み込みを開発環境とテスト環境のみに制限
+    if Rails.env.development? || Rails.env.test?
+      config.before_configuration do
+        env_file = File.join(Rails.root, 'config', 'local_env.yml')
+        if File.exist?(env_file)
+          YAML.load(File.open(env_file)).each do |key, value|
+            ENV[key.to_s] = value
+          end
+        end
+      end
     end
+
     # Configuration for the application, engines, and railties goes here.
     #
     # These settings can be overridden in specific environments using the files
